+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## Development
 
-* 2025-04-15 -- 0.2.0 Editor
+* 2025-08-17 -- 0.3.0 Progress Pages
+  * Added different progress for Asmr pagers
+  * Added notes to titles
+  * Delete confirmation
+  * Check for duplicate works
+  * Accept RJ Code or link
+  * Fixes
+
+* 2025-04-17 -- 0.2.0 Editor
   * Editor
   * CRUD